重複部分をcomposablesに抜き出し
