+++ src/components/form/CAutocomplete.vue
@@ -1,5 +1,7 @@
 <script setup lang="ts">
 import { computed, reactive,  ref,  watchEffect } from 'vue'
+import { getScrollParent } from '@/composables/scroll';
+import { teleportTarget } from '@/composables/teleport'
 import { mdiMenuDown, mdiMenuUp, mdiClose } from '@mdi/js';
 import CSvgIcon from '@/components/images/CSvgIcon.vue';
 
@@ -171,15 +173,6 @@ const selectionSlotDisplay = computed(() => {
     return props.itemValue ? Object.keys(selectionItem).length : selectionItem.value!==''
 })
 
-const teleportTarget = computed(() => {
-    if (typeof window === 'undefined') return undefined
-    const targetElement = document.body
-    const container = document.createElement('div')
-    container.className = 'c-overlay-container'
-    targetElement.appendChild(container)
-    return container
-})
-
 const liClass= (item:any) => {
     if(props.itemValue) return item[props.itemValue]==selectionItem.value[props.itemValue]?'bg-blue-50 text-blue-700':''
     return item == selectionItem.value ? 'bg-blue-50 text-blue-700':''
@@ -229,22 +222,6 @@ const optionClass = () => {
     else optionsPosition.top = fieldEl.value.getBoundingClientRect().top + fieldEl.value.clientHeight + 1 + 'px'
     optionsPosition.left = fieldEl.value.getBoundingClientRect().left+'px'
     optionsPosition.width = fieldEl.value.clientWidth+'px'
-}
-
-const getScrollParent = (el?: HTMLElement) => {
-    while (el) {
-        if (hasScrollbar(el)) return el
-        el = el.parentElement!
-    }
-
-    return document.scrollingElement as HTMLElement
-}
-
-const hasScrollbar = (el?: Element | null) => {
-    if (!el || el.nodeType !== Node.ELEMENT_NODE) return false
-
-    const style = window.getComputedStyle(el)
-    return style.overflowY === 'scroll' || (style.overflowY === 'auto' && el.scrollHeight > el.clientHeight)
 }
 
 watchEffect(() => {
